Clean up changes
Changed output file writing. The file is only written to if a key is changed. No need to write to the file if keys remain the same.
Fixed issue with cancel button colour not changing correctly. Sometimes it's colour would remain stuck on white instead of reverting back to red. 
Changed control of recording key presses. Previously if a key was selected and then the player pressed cancel sometimes the key wouldn't reset and if another key was pressed then the first key's key bind would sometimes be changed.